+++ sdk/dart/lib/cel_bridge.dart
@@ -7,6 +7,7 @@ export 'src/cel_exception.dart';
 export 'src/cel_issue.dart';
 export 'src/cel_request_result.dart';
 export 'src/cel_runtime.dart';
+export 'src/cel_runtime_limits.dart';
 export 'src/cel_runtime_options.dart';
 export 'src/cel_type.dart';
 export 'src/cel_validation_result.dart';